refactor(frontend): split audiobooks/podcasts/soulseek/enrichment/metadata api mixins
Move the audiobook, podcast, soulseek, enrichment, and metadata-editing
domain methods out of api.ts into typed mixin modules. Pure structural move.

Refactor slice P34 (api-client-typing-split).
