+++ frontend/lib/api.ts
@@ -1,16 +1,21 @@
 import { ApiClientCore, type ApiData } from "./api/core";
 import { WithAuth } from "./api/auth";
+import { WithAudiobooks } from "./api/audiobooks";
 import { WithConnectors } from "./api/connectors";
+import { WithEnrichment } from "./api/enrichment";
 import { WithDiscover } from "./api/discover";
 import { WithDownloads } from "./api/downloads";
 import { WithImports } from "./api/imports";
 import { WithLibrary } from "./api/library";
 import { WithMedia } from "./api/media";
+import { WithMetadata } from "./api/metadata";
 import { WithNotifications } from "./api/notifications";
 import { WithPlaylists } from "./api/playlists";
+import { WithPodcasts } from "./api/podcasts";
 import { WithPlays } from "./api/plays";
 import { WithRecommendations } from "./api/recommendations";
 import { WithSettings } from "./api/settings";
+import { WithSoulseek } from "./api/soulseek";
 import type {
     YouTubePlaylistInfo,
     YouTubeDownloadJob,
@@ -360,488 +365,17 @@ export interface PlaybackClientMetricInput {
     fields?: Record<string, unknown>;
 }
 
-class ApiClient extends WithNotifications(WithDiscover(WithImports(WithDownloads(WithAuth(
-    WithConnectors(WithSettings(WithPlays(WithRecommendations(WithMedia(
-        WithPlaylists(WithLibrary(ApiClientCore))
+class ApiClient extends WithAudiobooks(WithPodcasts(WithSoulseek(WithEnrichment(WithMetadata(
+    WithNotifications(WithDiscover(WithImports(WithDownloads(WithAuth(
+        WithConnectors(WithSettings(WithPlays(WithRecommendations(WithMedia(
+            WithPlaylists(WithLibrary(ApiClientCore))
+        )))))
     )))))
 ))))) {
     async getYtMusicMixes(): Promise<{
         mixes: Array<{ playlistId: string; title: string; description: string; thumbnails: Array<{ url: string; width: number }>; count: string | null }>;
     }> {
         return this.get("/browse/ytmusic/mixes");
-    }
-
-    // Audiobooks
-    async getAudiobooks() {
-        return this.request<ApiData[]>("/audiobooks");
-    }
-
-    async getAudiobook(id: string) {
-        return this.request<ApiData>(`/audiobooks/${id}`);
-    }
-
-    async getAudiobookSeries(seriesName: string) {
-        return this.request<ApiData[]>(
-            `/audiobooks/series/${encodeURIComponent(seriesName)}`
-        );
-    }
-
-    getAudiobookStreamUrl(id: string): string {
-        const baseUrl = `${this.getBaseUrl()}/api/audiobooks/${id}/stream`;
-        // For audio element requests, cookies may not be sent cross-origin in development
-        // Add token as query param for authentication (supported by requireAuthOrToken)
-        const token = this.getCurrentToken();
-        if (token) {
-            return `${baseUrl}?token=${encodeURIComponent(token)}`;
-        }
-        return baseUrl;
-    }
-
-    async updateAudiobookProgress(
-        id: string,
-        currentTime: number,
-        duration: number,
-        isFinished: boolean = false
-    ) {
-        return this.request<ApiData>(`/audiobooks/${id}/progress`, {
-            method: "POST",
-            body: JSON.stringify({ currentTime, duration, isFinished }),
-        });
-    }
-
-    async deleteAudiobookProgress(id: string) {
-        return this.request<ApiData>(`/audiobooks/${id}/progress`, {
-            method: "DELETE",
-        });
-    }
-
-    async getContinueListening() {
-        return this.request<ApiData[]>("/audiobooks/continue-listening");
-    }
-
-    async searchAudiobooks(query: string) {
-        return this.request<ApiData[]>(
-            `/audiobooks/search?q=${encodeURIComponent(query)}`
-        );
-    }
-
-    // Podcasts
-    async getPodcasts() {
-        return this.request<ApiData[]>("/podcasts");
-    }
-
-    async getPodcast(id: string) {
-        return this.request<ApiData>(`/podcasts/${id}`, { silent404: true });
-    }
-
-    async previewPodcast(itunesId: string) {
-        return this.request<ApiData>(`/podcasts/preview/${itunesId}`);
-    }
-
-    getPodcastEpisodeStreamUrl(podcastId: string, episodeId: string): string {
-        const baseUrl = `${this.getBaseUrl()}/api/podcasts/${podcastId}/episodes/${episodeId}/stream`;
-        // For audio element requests, cookies may not be sent cross-origin in development
-        // Add token as query param for authentication (supported by requireAuthOrToken)
-        const token = this.getCurrentToken();
-        if (token) {
-            return `${baseUrl}?token=${encodeURIComponent(token)}`;
-        }
-        return baseUrl;
-    }
-
-    /**
-     * Check if a podcast episode is cached locally
-     * Returns { cached: boolean, downloading: boolean, downloadProgress: number | null }
-     */
-    async getPodcastEpisodeCacheStatus(
-        podcastId: string,
-        episodeId: string
-    ): Promise<{
-        cached: boolean;
-        downloading: boolean;
-        downloadProgress: number | null;
-    }> {
-        return this.request<{
-            cached: boolean;
-            downloading: boolean;
-            downloadProgress: number | null;
-        }>(`/podcasts/${podcastId}/episodes/${episodeId}/cache-status`);
-    }
-
-    async updatePodcastEpisodeProgress(
-        podcastId: string,
-        episodeId: string,
-        currentTime: number,
-        duration: number,
-        isFinished: boolean = false
-    ) {
-        return this.request<ApiData>(
-            `/podcasts/${podcastId}/episodes/${episodeId}/progress`,
-            {
-                method: "POST",
-                body: JSON.stringify({ currentTime, duration, isFinished }),
-            }
-        );
-    }
-
-    // Alias for compatibility with AudioElement
-    async updatePodcastProgress(
-        podcastId: string,
-        episodeId: string,
-        currentTime: number,
-        duration: number,
-        isFinished: boolean = false
-    ) {
-        return this.updatePodcastEpisodeProgress(
-            podcastId,
-            episodeId,
-            currentTime,
-            duration,
-            isFinished
-        );
-    }
-
-    async deletePodcastEpisodeProgress(podcastId: string, episodeId: string) {
-        return this.request<ApiData>(
-            `/podcasts/${podcastId}/episodes/${episodeId}/progress`,
-            {
-                method: "DELETE",
-            }
-        );
-    }
-
-    async getSimilarPodcasts(podcastId: string) {
-        return this.request<ApiData[]>(`/podcasts/${podcastId}/similar`);
-    }
-
-    async getTopPodcasts(limit = 20, genreId?: number) {
-        const params = new URLSearchParams({ limit: limit.toString() });
-        if (genreId) params.append("genreId", genreId.toString());
-        return this.request<ApiData[]>(
-            `/podcasts/discover/top?${params.toString()}`
-        );
-    }
-
-    async getPodcastsByGenre(genreIds: number[]) {
-        return this.request<ApiData>(
-            `/podcasts/discover/genres?genres=${genreIds.join(",")}`
-        );
-    }
-
-    async getPodcastsByGenrePaginated(genreId: number, limit = 20, offset = 0) {
-        return this.request<ApiData[]>(
-            `/podcasts/discover/genre/${genreId}?limit=${limit}&offset=${offset}`
-        );
-    }
-
-    async subscribePodcast(feedUrl: string, itunesId?: string) {
-        return this.request<{ success: boolean; podcast?: ApiData }>("/podcasts/subscribe", {
-            method: "POST",
-            body: JSON.stringify({ feedUrl, itunesId }),
-        });
-    }
-
-    async removePodcast(podcastId: string) {
-        return this.request<{ success: boolean; message: string }>(
-            `/podcasts/${podcastId}/unsubscribe`,
-            {
-                method: "DELETE",
-            }
-        );
-    }
-
-    // Soulseek - P2P Music Search & Download
-    async getSlskdStatus() {
-        return this.request<{
-            enabled: boolean;
-            connected: boolean;
-            username?: string;
-            message?: string;
-        }>("/soulseek/status");
-    }
-
-    async searchSoulseek(query: string) {
-        return this.request<{ searchId: string; message: string }>(
-            "/soulseek/search",
-            {
-                method: "POST",
-                body: JSON.stringify({ query }),
-            }
-        );
-    }
-
-    async getSoulseekResults(searchId: string) {
-        return this.request<{ results: ApiData[]; count: number }>(
-            `/soulseek/search/${searchId}`
-        );
-    }
-
-    async downloadFromSoulseek(
-        username: string,
-        filepath: string,
-        filename?: string,
-        size?: number,
-        artist?: string,
-        album?: string,
-        title?: string
-    ) {
-        return this.request<{
-            success: boolean;
-            message: string;
-            filename: string;
-        }>("/soulseek/download", {
-            method: "POST",
-            body: JSON.stringify({
-                username,
-                filepath,
-                filename,
-                size,
-                artist,
-                album,
-                title,
-            }),
-        });
-    }
-
-    // Enrichment
-    async getEnrichmentSettings() {
-        return this.request<ApiData>("/enrichment/settings");
-    }
-
-    async updateEnrichmentSettings(settings: ApiData) {
-        return this.request<ApiData>("/enrichment/settings", {
-            method: "PUT",
-            body: JSON.stringify(settings),
-        });
-    }
-
-    async enrichArtist(artistId: string) {
-        return this.request<{
-            success: boolean;
-            confidence: number;
-            data: ApiData;
-        }>(`/enrichment/artist/${artistId}`, {
-            method: "POST",
-        });
-    }
-
-    async enrichAlbum(albumId: string) {
-        return this.request<{
-            success: boolean;
-            confidence: number;
-            data: ApiData;
-        }>(`/enrichment/album/${albumId}`, {
-            method: "POST",
-        });
-    }
-
-    async startLibraryEnrichment() {
-        return this.request<{ success: boolean; message: string }>(
-            "/enrichment/start",
-            {
-                method: "POST",
-            }
-        );
-    }
-
-    async syncLibraryEnrichment() {
-        return this.request<{
-            message: string;
-            description: string;
-            result: {
-                artists: number;
-                tracks: number;
-                audioQueued: number;
-            };
-        }>("/enrichment/sync", {
-            method: "POST",
-        });
-    }
-
-    async getEnrichmentProgress() {
-        return this.request<{
-            artists: {
-                total: number;
-                completed: number;
-                pending: number;
-                failed: number;
-                progress: number;
-            };
-            trackTags: {
-                total: number;
-                enriched: number;
-                pending: number;
-                progress: number;
-            };
-            audioAnalysis: {
-                total: number;
-                completed: number;
-                pending: number;
-                processing: number;
-                failed: number;
-                progress: number;
-                isBackground: boolean;
-            };
-            clapEmbeddings: {
-                total: number;
-                completed: number;
-                pending: number;
-                processing: number;
-                failed: number;
-                progress: number;
-                isBackground: boolean;
-            };
-            coreComplete: boolean;
-            isFullyComplete: boolean;
-        }>("/enrichment/progress");
-    }
-
-    async triggerFullEnrichment(options?: {
-        forceVibeRebuild?: boolean;
-        forceMoodBucketBackfill?: boolean;
-    }) {
-        return this.request<{
-            message: string;
-            description: string;
-            forceVibeRebuild?: boolean;
-            forceMoodBucketBackfill?: boolean;
-        }>(
-            "/enrichment/full",
-            {
-                method: "POST",
-                body: JSON.stringify({
-                    forceVibeRebuild: options?.forceVibeRebuild === true,
-                    forceMoodBucketBackfill:
-                        options?.forceMoodBucketBackfill === true,
-                }),
-            }
-        );
-    }
-
-    async resetArtistsOnly() {
-        return this.request<{
-            message: string;
-            description: string;
-            count: number;
-        }>("/enrichment/reset-artists", { method: "POST" });
-    }
-
-    async resetMoodTagsOnly() {
-        return this.request<{
-            message: string;
-            description: string;
-            count: number;
-        }>("/enrichment/reset-mood-tags", { method: "POST" });
-    }
-
-    async resetAudioAnalysisOnly() {
-        return this.request<{
-            message: string;
-            description: string;
-            count: number;
-        }>("/enrichment/reset-audio-analysis", { method: "POST" });
-    }
-
-    async retryFailedAnalysis() {
-        return this.request<{ message: string; reset: number }>("/analysis/retry-failed", {
-            method: "POST",
-        });
-    }
-
-    async updateArtistMetadata(
-        artistId: string,
-        data: {
-            name?: string;
-            bio?: string;
-            genres?: string[];
-            mbid?: string;
-            heroUrl?: string;
-        }
-    ) {
-        return this.request<ApiData>(`/enrichment/artists/${artistId}/metadata`, {
-            method: "PUT",
-            body: JSON.stringify(data),
-        });
-    }
-
-    async updateAlbumMetadata(
-        albumId: string,
-        data: {
-            title?: string;
-            year?: number;
-            genres?: string[];
-            rgMbid?: string;
-            coverUrl?: string;
-        }
-    ) {
-        return this.request<ApiData>(`/enrichment/albums/${albumId}/metadata`, {
-            method: "PUT",
-            body: JSON.stringify(data),
-        });
-    }
-
-    async updateTrackMetadata(trackId: string, data: ApiData) {
-        return this.request<ApiData>(`/enrichment/tracks/${trackId}/metadata`, {
-            method: "PUT",
-            body: JSON.stringify(data),
-        });
-    }
-
-    async resetArtistMetadata(artistId: string) {
-        return this.request<{ message: string; artist: ApiData }>(
-            `/enrichment/artists/${artistId}/reset`,
-            { method: "POST" }
-        );
-    }
-
-    async resetAlbumMetadata(albumId: string) {
-        return this.request<{ message: string; album: ApiData }>(
-            `/enrichment/albums/${albumId}/reset`,
-            { method: "POST" }
-        );
-    }
-
-    async resetTrackMetadata(trackId: string) {
-        return this.request<{ message: string; track: ApiData }>(
-            `/enrichment/tracks/${trackId}/reset`,
-            { method: "POST" }
-        );
-    }
-
-    async searchMusicBrainzArtists(query: string): Promise<{
-        artists: Array<{
-            mbid: string;
-            name: string;
-            disambiguation: string | null;
-            country: string | null;
-            type: string | null;
-            score: number;
-        }>;
-    }> {
-        return this.request(
-            `/enrichment/search/musicbrainz/artists?q=${encodeURIComponent(query)}`
-        );
-    }
-
-    async searchMusicBrainzReleaseGroups(
-        query: string,
-        artistName?: string
-    ): Promise<{
-        albums: Array<{
-            rgMbid: string;
-            title: string;
-            primaryType: string;
-            secondaryTypes: string[];
-            firstReleaseDate: string | null;
-            artistCredit: string;
-            score: number;
-        }>;
-    }> {
-        let url = `/enrichment/search/musicbrainz/release-groups?q=${encodeURIComponent(query)}`;
-        if (artistName) {
-            url += `&artist=${encodeURIComponent(artistName)}`;
-        }
-        return this.request(url);
     }
 
     // Vibe (CLAP Similarity) API
@@ -1024,21 +558,6 @@ class ApiClient extends WithNotifications(WithDiscover(WithImports(WithDownloads
             updatedAt?: string;
             quantiles: number[];
         }>("/vibe/calibration");
-    }
-
-    async refreshAllPodcasts() {
-        return this.request<{
-            success: boolean;
-            total: number;
-            totalNewEpisodes: number;
-            failed: number;
-            results: Array<{
-                podcastId: string;
-                success: boolean;
-                newEpisodesCount: number;
-                error?: string;
-            }>;
-        }>("/podcasts/refresh-all", { method: "POST" });
     }
 
     // ── YouTube Music ──────────────────────────────────────────────
