+++ frontend/lib/api/audiobooks.ts
@@ -1,0 +1,62 @@
+import { type ApiClientConstructor, type ApiData } from "./core";
+
+/** Add audiobook-domain operations to an API client base class. */
+export function WithAudiobooks<TBase extends ApiClientConstructor>(Base: TBase) {
+    abstract class AudiobooksApi extends Base {
+
+    // Audiobooks
+    async getAudiobooks() {
+        return this.request<ApiData[]>("/audiobooks");
+    }
+
+    async getAudiobook(id: string) {
+        return this.request<ApiData>(`/audiobooks/${id}`);
+    }
+
+    async getAudiobookSeries(seriesName: string) {
+        return this.request<ApiData[]>(
+            `/audiobooks/series/${encodeURIComponent(seriesName)}`
+        );
+    }
+
+    getAudiobookStreamUrl(id: string): string {
+        const baseUrl = `${this.getBaseUrl()}/api/audiobooks/${id}/stream`;
+        // For audio element requests, cookies may not be sent cross-origin in development
+        // Add token as query param for authentication (supported by requireAuthOrToken)
+        const token = this.getCurrentToken();
+        if (token) {
+            return `${baseUrl}?token=${encodeURIComponent(token)}`;
+        }
+        return baseUrl;
+    }
+
+    async updateAudiobookProgress(
+        id: string,
+        currentTime: number,
+        duration: number,
+        isFinished: boolean = false
+    ) {
+        return this.request<ApiData>(`/audiobooks/${id}/progress`, {
+            method: "POST",
+            body: JSON.stringify({ currentTime, duration, isFinished }),
+        });
+    }
+
+    async deleteAudiobookProgress(id: string) {
+        return this.request<ApiData>(`/audiobooks/${id}/progress`, {
+            method: "DELETE",
+        });
+    }
+
+    async getContinueListening() {
+        return this.request<ApiData[]>("/audiobooks/continue-listening");
+    }
+
+    async searchAudiobooks(query: string) {
+        return this.request<ApiData[]>(
+            `/audiobooks/search?q=${encodeURIComponent(query)}`
+        );
+    }
+    }
+    return AudiobooksApi;
+}
